Подкорректированы модели, enum вынесены в отдельный файл
